refactor: limpando modelos
