--- src/modelos/Senha.go
@@ -1,7 +1,0 @@
-package modelos
-
-// Struct Senha representa o formato da requisição de alteração de senha
-type Senha struct {
-	Nova  string `json:"nova"`
-	Atual string `json:"atual"`
-}
